Moved modules into own package, using relative imports. Created starter script.
