+++ g.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+PYTHONPATH=. /usr/bin/env python3 -m gruschen.main $*
