+++ gruschen/signal_type.py
