+++ gruschen/sound_file.py
@@ -5,7 +5,7 @@ import wave
 
 import numpy as np
 
-from signal_type import Signal
+from .signal_type import Signal
 
 
 class SoundFileTest(unittest.TestCase):
