+++ gruschen/trimming.py
@@ -1,5 +1,5 @@
 
-from signal_type import Signal
+from .signal_type import Signal
 
 
 def trim_all_zeros(signal):
