+++ gruschen/utility.py
